+++ src/main/java/com/parkmate/notificationservice/notificationsender/firebase/FireBaseInitializer.java
@@ -1,4 +1,4 @@
-package com.parkmate.notificationservice.firebase;
+package com.parkmate.notificationservice.notificationsender.firebase;
 
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
